Prefer borrows over clones; close an rm race; widen ids to i64
Borrow-first cleanups across the workspace, plus two correctness fixes
found along the way.

Matcher hits could not be kept. `for_each_hit`'s elided lifetime was
higher-ranked, so its documented contract ("visitors clone only what they
keep") was unachievable: nothing could be kept, and every visitor was
forced to clone. Binding `&'a self` fixes it — hits borrow from
`self.groups`, never from the haystack, so `text` stays unconstrained.
`scan` and `suggest-o2m` now aggregate into maps keyed by `&str` /
`&PatternTarget`, allocating once per distinct result instead of twice per
hit on a path whose own docs warn that hits can far outnumber the text. A
regression test fails to compile if the lifetime is ever re-elided.

Chunked inserts cloned every row. Six sites passed
`chunks(..).iter().cloned()` to `insert_many`, copying each `ActiveModel`
on its way into SQL. One `insert_chunked` helper drains instead, covering
every write path.

`rm` had a check-then-delete race. It ran get -> descendant probe ->
delete as three statements, so a concurrent `add` could orphan a
descendant — the same hazard `delete_subtree` already documents solving.
`BranchWriter::delete_leaf` now does all three in one transaction, and
`rm` narrows to a pure write.

Five hand-rolled pagination loops became one `PageCursor`. A cursor rather
than a `for_each_page(callback)`: `stats` must await per page, which a
synchronous closure cannot do.

Row ids are `i64`. Migrations already use `INTEGER PRIMARY KEY
AUTOINCREMENT` (64-bit), so the `i32` entities imposed a fallible
narrowing for no storage benefit. `sqlite_branch_id` is gone and
`branch_active_model` is infallible. Behavior change: an unknown branch id
is now refused by the foreign key rather than by a width check, which is
the stronger guarantee; three tests were rewritten to assert it.

`Branch` equality now compares the domain value. The derived `PartialEq`
folded in `id`/`revision`, which are `#[serde(skip)]`, so a stored branch
never equalled the same branch parsed from a file — exactly the comparison
import/export tests make. Removes a test workaround that nulled both
fields by hand.

Also: `SampleQuery::from_slot` consumes its slot and `resolve_query_slot`
takes the slot out of the frame (this removed a whole-`Slot` clone plus
six map clones); `Selection.slot` is an `Rc<str>`;
`save_packets_with_usages` consumes records that each hold a serialized
packet; `ls` uses `strip_prefix` instead of a panic-capable byte slice;
subqueries bind values by borrow.

`commands.rs` (819 lines) splits into `commands/{branch,navigate,frames}`
and `usage` splits from `sampling`.

fmt, clippy -D warnings, rustdoc -D warnings, 154 tests, release build and
cargo package all pass. cargo-deny was not run: it is not installed here,
and no dependencies changed.

+++ crates/unclip-entity/src/entities/branch_o2m_values.rs
@@ -7,7 +7,7 @@ use serde::{Deserialize, Serialize};
 #[sea_orm(table_name = "branch_o2m_values")]
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
-    pub branch_id: i32,
+    pub branch_id: i64,
     #[sea_orm(primary_key, auto_increment = false, column_type = "Text")]
     pub name: String,
     #[sea_orm(primary_key, auto_increment = false, column_type = "Text")]
